refactor: extract and apply theme to Plotly layouts for improved consistency in hyperparameters plots and explainers
