+++ DashAI/front/src/components/explainers/ExplainersPlot.jsx
@@ -14,82 +14,7 @@ import { useSnackbar } from "notistack";
 
 import { getExplainerPlot as getExplainerPlotRequest } from "../../api/explainer";
 import { useTranslation } from "react-i18next";
-
-function applyThemeToLayout(baseLayout, theme) {
-  const bg = theme.palette.background.paper;
-  const textColor = theme.palette.text.primary;
-  const gridColor = theme.palette.divider;
-
-  // Strip backend-provided fixed dimensions so our responsive sizing takes over
-  // eslint-disable-next-line no-unused-vars
-  const { width: _w, height: _h, ...rest } = baseLayout ?? {};
-
-  const axisOverride = {
-    gridcolor: gridColor,
-    zerolinecolor: gridColor,
-  };
-
-  return {
-    ...rest,
-    paper_bgcolor: bg,
-    plot_bgcolor: bg,
-    font: {
-      ...baseLayout?.font,
-      color: textColor,
-      family: "Quicksand-Bold, sans-serif",
-    },
-    xaxis: {
-      ...baseLayout?.xaxis,
-      ...axisOverride,
-      tickfont: { ...baseLayout?.xaxis?.tickfont, color: textColor },
-    },
-    yaxis: {
-      ...baseLayout?.yaxis,
-      ...axisOverride,
-      tickfont: { ...baseLayout?.yaxis?.tickfont, color: textColor },
-    },
-    legend: {
-      ...baseLayout?.legend,
-      bgcolor: bg,
-      bordercolor: gridColor,
-    },
-    title: {
-      ...baseLayout?.title,
-      font: { ...baseLayout?.title?.font, color: textColor },
-    },
-    ...(baseLayout?.updatemenus && {
-      updatemenus: baseLayout.updatemenus.map((menu) => ({
-        ...menu,
-        bgcolor:
-          theme.palette.ui?.borderLight ?? theme.palette.background.paper,
-        bordercolor: gridColor,
-        font: { color: "#000000" },
-        activecolor: theme.palette.primary.main,
-      })),
-    }),
-    ...(baseLayout?.polar && {
-      polar: {
-        ...baseLayout.polar,
-        bgcolor: bg,
-        radialaxis: {
-          ...baseLayout.polar.radialaxis,
-          gridcolor: gridColor,
-          linecolor: gridColor,
-          tickfont: {
-            ...baseLayout.polar.radialaxis?.tickfont,
-            color: textColor,
-          },
-        },
-        angularaxis: {
-          ...baseLayout.polar.angularaxis,
-          color: textColor,
-          gridcolor: gridColor,
-          linecolor: gridColor,
-        },
-      },
-    }),
-  };
-}
+import { applyThemeToLayout } from "../../utils/plotlyTheme";
 
 export default function ExplainersPlot({ explainer, scope }) {
   const { enqueueSnackbar } = useSnackbar();
